Refactor: Move catalog_types import from utils to topic_handlers across multiple files
